% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comcode_description.R
\name{comcode_description}
\alias{comcode_description}
\title{Hierarchichal classification of a given commodity code}
\usage{
comcode_description(code, lookup_table)
}
\arguments{
\item{code}{A character string representing any HS2/HS4/HS6/CN8 code.}

\item{lookup_table}{A `data.table` containing details of each code level.}
}
\value{
A display of the hierarchical information of code.
}
\description{
This function displays the information related to a given commodity code and
any of its subcodes (HS2/HS4/HS6/CN8).
}
