% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_ts}
\alias{extract_ts}
\title{Extract of monthly or daily time series for a given commodity code}
\usage{
extract_ts(
  data,
  code,
  date_col = "DATE_START",
  response_col = "NET_MASS",
  fill_missing = NA,
  freq = NULL,
  group_by = NULL,
  return_list = FALSE
)
}
\arguments{
\item{data}{A `data.table` containing trade data. Must include columns
`COMCODE` and specified `quantity`.}

\item{code}{Any single or groups of HS2/HS4/HS6/CN8 code.}

\item{date_col}{Name of column containing timestamps.}

\item{response_col}{Quantity to be extracted and aggregated as time series, e.g.
`NET_MASS` or `STAT_VALUE` or `volume`. NA are omitted when summing values.}

\item{fill_missing}{This function returns a continuous time series. Values
for missing dates are filled with this value.}

\item{freq}{Frequency of time series data. See details.}

\item{group_by}{Optional vector of column names used to create
separate time series. For example, use `"PORT_CODE"` to return one series
per port, or `c("PORT_CODE", "COO_ALPHA")` for one series per port-country of origin
combination. Default set to `NULL`, which returns time series for the selected
commodity code.}

\item{return_list}{If `TRUE`, returns a list of completed time
series per code and group. If `FALSE`, returns one combined `data.table`.}
}
\value{
A `data.table` with date and quantity columns.
}
\description{
This function extracts the sum of a given quantity from import_data for
any level of hierarchy (HS2, HS4, HS6 and CN8) found in the dataset.
}
\details{
Daily or monthly data is expected and detected automatically. Missing
values are filled.

By default, `freq=NULL` and will be detected automatically. However, this may
fail for sparse data, in which case `freq` should be set manually. Options are
"day", "week", or "month".
}
