% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_volume_outliers.R
\name{monthly_volume_outliers}
\alias{monthly_volume_outliers}
\title{Display of outliers from residuals for a given commodity code}
\usage{
monthly_volume_outliers(import_data, code, lookup_table)
}
\arguments{
\item{import_data}{A `data.table` containing trade data. Must include "COMCODE".}

\item{code}{A character string representing any HS2/HS4/HS6/CN8 code.}

\item{lookup_table}{A `data.table` containing details of each code level.}
}
\value{
A scatter plot of fitted values against residuals and 
a display of residuals over time per month per code.
}
\description{
This function displays outliers in residuals of import volume after estimating 
an ARX model for any level of hierarchy (HS2, HS4, HS6 and CN8) found in the dataset.
}
