% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{open_userguide}
\alias{open_userguide}
\title{View User Guide}
\usage{
open_userguide(path = NULL)
}
\arguments{
\item{path}{Optional. Path to specific instance of `UserGuide.html`. If `NULL`, it will be retrieved from the current installation of `bulktrends`.}
}
\description{
A function that opens the `bulktrends` user guide in system browser.
}
