% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_best_model.R
\name{select_best_model}
\alias{select_best_model}
\title{Extract suitable model matrix for a given commodity code}
\usage{
select_best_model(
  data,
  date_col = "DATE_START",
  formulas = NULL,
  response_col = NULL,
  metric = "aic",
  scale_ts = FALSE,
  freq = NULL
)
}
\arguments{
\item{data}{A `data.frame` containing the dependent variable and dates.}

\item{date_col}{Name of column containing timestamps.}

\item{formulas}{A list of formulas specifying candidate models. Covariates
available are `linear_trend` and `month`.}

\item{response_col}{If `formulas` not provided, default formulas will be used
but the column containing the response variable must be specified.}

\item{metric}{A character string specifying the criteria for model
selection. Examples are "aic","aicc" or "bic".}

\item{scale_ts}{If `TRUE`, time series is scaled to zero mean and unit variance using `scale()`. Default `FALSE`.}

\item{freq}{Frequency of the input time series. If NULL it will be auto-detected ('day', 'week', or 'month').}
}
\value{
A model matrix of the linear_trend and seasonal regressors of the selected
model and the related model formula.
}
\description{
This function evaluates a set of formulas with linear_trend and seasonal terms
as exogenous variables and selects the ARIMA model with the lowest value
of the selected metric.
}
