% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{uktrade_tsibble}
\alias{uktrade_tsibble}
\title{Convert UK trade data table to a hierarchical tsibble}
\usage{
uktrade_tsibble(
  data,
  response_col = "NET_MASS",
  comcode_level = c("HS2", "HS4"),
  group_by = NULL,
  date_col = "DATE_START",
  freq = NULL,
  fill_missing = NA
)
}
\arguments{
\item{data}{A `data.table` containing trade data. Must include columns
`COMCODE` and specified `response_col`.}

\item{response_col}{Quantity to be extracted and aggregated as time series, e.g.`NET_MASS`, `volume`}

\item{comcode_level}{Hierarchy columns to use as tsibble keys
and in hierarchical aggregation. Defaults to `c("HS2", "HS4")`.
Expected commodity hierarchy columns are `"HS2"`, `"HS4"`,`"HS6"`, `"CN8"`, and `"CN10"`.}

\item{group_by}{Optional classification group in `data`, e.g. `PORT_CODE`. Use `"category"` to
group commodities by plants and animals. Default is `NULL`.}

\item{date_col}{Name of column containing timestamps.}

\item{freq}{Frequency of time series data. See details.}

\item{fill_missing}{This function returns a continuous time series. Values
for missing dates are filled with this value.}
}
\value{
A hierarchical tsibble with date, quantity and aggregated hierarchy levels columns.
}
\description{
This function creates a tsibble from UK trade data by deriving commodity-code hierarchy
columns, aggregating a selected quantity over time, filling missing time
periods with zero, and aggregating values per hierarchies.
}
\details{
Daily or monthly data is expected and detected automatically. Missing
values are filled.

By default, `freq=NULL` and will be detected automatically. However, this may
fail for sparse data, in which case `freq` should be set manually. Options are
"day", "week", or "month".

The `category` column labels commodity codes found in
`comcode_groups$animal_SPS` as `"Animal"` and those in
`comcode_groups$plant_SPS` as `"Plant"`.
}
