% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uktradeinfo_api.R
\name{uktrades_request}
\alias{uktrades_request}
\title{Request data from UK Trades API}
\usage{
uktrades_request(endpoint)
}
\arguments{
\item{endpoint}{OData queryable endpoint appended to api URL. See details.}
}
\value{
Result of API query.
}
\description{
Request data from UK Trades API
}
\details{
For endpoint documentation visit https://www.uktradeinfo.com/api-documentation/.
`endpoint` is appended to "https://api.uktradeinfo.com/".
}
