% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_date_features}
\alias{add_date_features}
\title{Add date features}
\usage{
add_date_features(data, date_col)
}
\arguments{
\item{data}{A `data.table` containing trade data.}

\item{date_col}{Name of column containing timestamps.}
}
\value{
A `data.table` with the original data and additional calendar features, including
day of week, day of year, UK holiday indicators etc.
}
\description{
A function that identifies the features (bank holidays) of dates provided.
}
