% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{comcode_groups}
\alias{comcode_groups}
\title{COMCODE Groups}
\format{
## `comcode_groups`
A named `list` containing a groups of commodity codes
\describe{
  \item{name}{Name of group}
  \item{value}{vector of Cn8 codes (`charater`) in specified group}
}
}
\usage{
comcode_groups
}
\description{
Useful groups of commodity codes for filtering import data.
}
\keyword{datasets}
