% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comcode_plot.R
\name{comcode_plot}
\alias{comcode_plot}
\title{Visualisation of a given commodity code.
This function displays the total weight per month for all 
levels of hierarchy (HS2, HS4, HS6 and CN8) found in the dataset.}
\usage{
comcode_plot(import_data, code, lookup_table, max_unique_comcodes = 10)
}
\arguments{
\item{import_data}{A `data.table` containing trade data. Must include "COMCODE".}

\item{code}{A character string representing any HS2/HS4/HS6/CN8 code.}

\item{lookup_table}{A `data.table` containing details of each code level.}
}
\value{
A display of total weight per month per code.
}
\description{
Visualisation of a given commodity code.
This function displays the total weight per month for all 
levels of hierarchy (HS2, HS4, HS6 and CN8) found in the dataset.
}
