% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_anomalies.R
\name{detect_anomalies}
\alias{detect_anomalies}
\title{Large-scale anomaly detection}
\usage{
detect_anomalies(
  import_data,
  codes,
  quantity = "NET_MASS",
  date_col = "DATE_START",
  model_selection_metric = "aic",
  scale_ts = FALSE,
  freq = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{import_data}{A `data.table` containing trade data. Must include columns `COMCODE` and specified `quantity`.}

\item{codes}{A vector of HS2/HS4/HS6/CN8 codes}

\item{quantity}{Quantity to be analysed, e.g. `NET_MASS`, `STAT_VALUE` or `volume`.}

\item{date_col}{Name of column containing timestamps.}

\item{model_selection_metric}{Selection criteria passed to `select_best_model()`}

\item{scale_ts}{If `TRUE`, time series is scaled to zero mean and unit variance using `scale()`. Default `FALSE`.}

\item{freq}{See `?extract_ts()`}

\item{verbose}{If `TRUE`, progress messages are displayed. Default `FALSE`.}

\item{...}{Additional arguments passed to `tso()`}
}
\value{
A list containing a 1. table of detected outliers and 2. a list
of time series data with regressors, including outlier effects.
}
\description{
This function loops over a list of commodity codes, selects a regression model,
and detects anomalies.
}
