% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_anomaly_netmass.R
\name{detect_anomalies}
\alias{detect_anomalies}
\title{Large scale anomaly detection}
\usage{
detect_anomalies(
  import_data,
  codes,
  quantity = "NET_MASS",
  model_selection_metric = "aic",
  cval = 5,
  types = c("AO", "LS", "TC", "IO"),
  ...
)
}
\arguments{
\item{import_data}{A `data.table` containing trade data. Must include columns "COMCODE", "month" and `quantity`.}

\item{codes}{A vector of HS2/HS4/HS6/CN8 codes}

\item{quantity}{Quantity to be analysed, e.g. "NET_MASS" or "STAT_VALUE".}

\item{model_selection_metric}{Selection criteria passed to `select_best_model()`}

\item{types}{See `?tso`}

\item{...}{Additional arguments passed to `tso()`}

\item{cavl}{See `?tso`}
}
\value{
A table of detected outliers.
}
\description{
This function loops over a list of commodity codes, selects a regression model,
and detects anomalies.
}
