% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_breaks.R
\name{detect_breaks}
\alias{detect_breaks}
\title{Detect structural breaks in a time series}
\usage{
detect_breaks(data, date_col = "DATE_START", formula)
}
\arguments{
\item{data}{A `data.frame` containing the dependent variable and dates.}

\item{date_col}{Name of column containing timestamps. Default `"DATE_START"`.}

\item{formula}{A formula specifying the regression model for break detection.}
}
\value{
A list containing (i) `break_entry`, a `data.table` with identified breakpoints and
their respective timing and, (ii) `segments`, a matrix of segment indicator
variable for use as a regressor.
Returns `NULL` if no breaks are detected.
}
\description{
This function uses [strucchangeRcpp::breakpoints()] with BIC-based selection to identify
structural breaks in a given time series.
}
