% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_date_frequency}
\alias{detect_date_frequency}
\title{Detect date frequency}
\usage{
detect_date_frequency(dates)
}
\arguments{
\item{dates}{A vector of dates or name of column containing timestamps.}
}
\value{
A character string indicating the detected date frequency: "day", "week", or "month".
}
\description{
This function detects the frequency of a time series.
}
