% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_outliers.R
\name{detect_outliers}
\alias{detect_outliers}
\title{Detection of temporary time series outliers}
\usage{
detect_outliers(
  data,
  quantity,
  scale_ts = FALSE,
  xreg = NULL,
  return_tso = FALSE,
  tso_params = list()
)
}
\arguments{
\item{data}{A `data.frame` containing the dependent variable and dates.}

\item{quantity}{Name of the column containing the time series values,
e.g. `"NET_MASS"`, `"STAT_VALUE"` or `volume`.}

\item{scale_ts}{If `TRUE`, time series is scaled to zero mean and unit variance using `scale()`. Default `FALSE`.}

\item{xreg}{matrix of external regressors passed to `tso()`. Default `NULL`.}

\item{return_tso}{If `TRUE`, the full `tso()` output is included in the
returned list as element `tso`. Default `FALSE`.}

\item{tso_params}{A named list of additional arguments passed directly to `tsoutliers::tso()`,
e.g. `list(cval = 5, types = c("AO", "TC", "IO"))`. Default `list()`.}
}
\value{
A list with elements (i) `data`, the original time series dataset with outlier covariates
columns appended when detected, and (ii) `outliers`, the outlier summary table returned by `tso()`.
If `return_tso = TRUE`, a third element `tso` is also included. Returns
  `NULL` if `tso()` errors.
}
\description{
This function detects temporary outliers in a given time series
using `tsoutliers::tso()`.
}
