% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_tso_anomalies.R
\name{detect_outliers}
\alias{detect_outliers}
\title{Detection of temporary time series outliers}
\usage{
detect_outliers(
  data,
  date_col = "DATE_START",
  code,
  quantity,
  model_formula,
  types = c("AO", "LS", "TC", "IO"),
  scale_ts = FALSE,
  xreg = NULL,
  ...
)
}
\arguments{
\item{data}{A `data.frame` containing the dependent variable and dates.}

\item{date_col}{Name of column containing timestamps.}

\item{code}{A vector of HS2/HS4/HS6/CN8 codes}

\item{quantity}{Quantity to be analysed, e.g. `NET_MASS`, `STAT_VALUE` or `volume`.}

\item{model_formula}{A formula specifying the regression model for outlier detection.}

\item{types}{Outlier types to detect. Typical inputs are: AO (Additive Outlier), TC (Temporary Change),
IO (Innovational Outlier), LS (Level Shift), or SLS(Seasonal Level Shift).}

\item{scale_ts}{If `TRUE`, time series is scaled to zero mean and unit variance using `scale()`. Default `FALSE`.}

\item{xreg}{matrix of external regressors passed to `tso()`.}

\item{...}{Additional arguments passed directly to `tsoutliers::tso()`.}
}
\value{
A list of (i) the original time series dataset with detected outlier covariates
columns appended when available, and (ii) the outlier summary table returned by `tso()`.
}
\description{
This function detects temporary outliers in a given time series
using `tsoutliers::tso()`.
}
