% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_ts}
\alias{extract_ts}
\title{Utility helper functions}
\usage{
extract_ts(import_data, code, quantity = "NET_MASS")
}
\arguments{
\item{import_data}{A `data.table` containing trade data. Must include columns "COMCODE", "month" and `quantity`.}

\item{code}{A character string representing any HS2/HS4/HS6/CN8 code.}

\item{quantity}{Quantity to be extracted and aggregated as time series, e.g. "NET_MASS" or "STAT_VALUE".}
}
\value{
Time series object of NET_MASS
}
\description{
Extract of monthly time series of "NET_MASS" for a given commodity code
}
\details{
This function extracts the monthly sum "NET_MASS" from import_data for
any level of hierarchy (HS2, HS4, HS6 and CN8) found in the dataset.
}
