% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ipaffs_data.R
\name{read_ipaffs}
\alias{read_ipaffs}
\title{Load bulk daily data from IPAFFS}
\usage{
read_ipaffs(path)
}
\description{
Load bulk daily data from IPAFFS
}
\details{
Leading pairs of 00s dropped
Missing leading 0 added to COMCODE of odd length
}
