% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_uktradeinfo_data.R
\name{read_uktradeinfo}
\alias{read_uktradeinfo}
\title{Load bulk monthly data from UK Trade Info}
\usage{
read_uktradeinfo(path)
}
\arguments{
\item{path}{Path to a `.txt` file, a `.zip` archive, or a directory
containing `.txt` and/or `.zip` files. Directories are searched
recursively.}
}
\value{
A `data.table` of trade data with columns containing timestamps.
}
\description{
Load a single bulk data file, a `.zip` archive, or all files in a directory
(and its sub-directories) from UK Trade Info.
}
\details{
`.zip` archives are extracted to a temporary directory, read, and the
temporary directory is deleted before the function returns.

This function can be slow to read a large number of files but is much faster
if a `future::plan()` is set that enables parallel computation.
}
