% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_uktradeinfo_data.R
\name{read_uktradeinfo}
\alias{read_uktradeinfo}
\title{Load bulk data form UK Trade Info}
\usage{
read_uktradeinfo(path)
}
\arguments{
\item{path}{Path to a file to read, or directory containing multiple files,
subdirectories are recursed.}
}
\value{
A `data.table` of trade data.
}
\description{
Load a single bulk data file of all files in a directory (and its
subdirectories) from UK Trade Info.
}
