% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_best_model.R
\name{select_best_model}
\alias{select_best_model}
\title{Select the best regression model for a time series}
\usage{
select_best_model(
  data,
  date_col = "DATE_START",
  formulas = NULL,
  response_col = NULL,
  metric = AIC,
  break_detection = TRUE,
  freq = NULL
)
}
\arguments{
\item{data}{A `data.frame` containing the dependent variable and dates.}

\item{date_col}{Name of column containing timestamps.}

\item{formulas}{A list of formulas specifying candidate models. Covariates
available are `linear_trend` and `month`.}

\item{response_col}{If `formulas` not provided, default formulas will be used
but the column containing the response variable must be specified.}

\item{metric}{A character string specifying the criteria for model
selection. Examples are "aic","aicc" or "bic".}

\item{break_detection}{If `TRUE`, structural break detection is
performed and break-adjusted models are evaluated.}

\item{freq}{Frequency of the input time series. If NULL it will be auto-detected ('day', 'week', or 'month').}
}
\value{
@return A list containing (i) the time series' data with any generated regressors
and structural break segment variables, (ii) the optimal selected model formula, and
(iii) a `data.table` describing any detected structural breaks or `NULL` if no breaks were selected
}
\description{
This function evaluates a set of candidate regression formulas containing
trend and seasonal regressors and selects the model with the lowest value
of the specified selection metric (for example AIC or BIC).
}
