% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_best_model.R
\name{select_best_model}
\alias{select_best_model}
\title{Extract suitable model matrix for a given commodity code}
\usage{
select_best_model(
  data,
  metric = "aic",
  formulas = list(~-1, ~1, ~linear_trend, ~sin(2 * pi * month/12) + cos(2 * pi *
    month/12), ~linear_trend + sin(2 * pi * month/12) + cos(2 * pi * month/12))
)
}
\arguments{
\item{data}{A time series object containing the dependent variable.}

\item{metric}{A character string specifying the criteria for model
selection. Examples are "aic","aicc" or "bic".}
}
\value{
A model matrix of the linear_trend and seasonal regressors of the selected
model and the related model formula.
}
\description{
This function evaluates a set of formulas with linear_trend and seasonal terms
as exogenous variables and selects the ARIMA model with the lowest value
of the selected metric.
}
