% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade-tariff_api.R
\name{tradetariff_request}
\alias{tradetariff_request}
\title{Request data from GOV.UK Trade Tariff API}
\usage{
tradetariff_request(endpoint)
}
\arguments{
\item{endpoint}{Endpoint appended to the base API URL. See details.}
}
\value{
Result of API query.
}
\description{
Request data from GOV.UK Trade Tariff API
}
\details{
For endpoint documentation visit https://docs.trade-tariff.service.gov.uk/.
`endpoint` is appended to "https://www.trade-tariff.service.gov.uk/api/v2/".
}
